pyproject.toml
[tool.poetry]
name = "personal-analytics-hub"
version = "0.1.0"
description = ""
authors = ["mishamsk <mishamsk@gmail.com>"]
packages = [{ include = "loader", from = "etl" }]

[tool.poetry.dependencies]
python = ">=3.10,<=3.11"
zeep = "^4.1.0"
click = "^8.1.3"
dbt-postgres = "^1.3.0"
drebedengi-python-api = "^0.2.0"
psycopg2 = "^2.9.4"
SQLAlchemy = { extras = ["mypy"], version = "^1.4.41" }
python-dotenv = "^0.20.0"
alembic = "^1.8.1"
PyYAML = "^6.0"
dataclasses-json = "^0.5.7"
nerddiary = "^0.3.2"
requests = "^2.28.1"
colorama = ">=0.3.9,<0.4.6"
rich = "^12.6.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^22.8.0"
mypy = "^0.971"
pre-commit = "^2.20.0"
flake8 = "^5.0.4"
bump2version = "^1.0.1"
tox = "^3.25.1"
ipython = "^7.30.1"
ipykernel = "^6.6.0"
types-PyYAML = "^6.0.11"
types-backports = "^0.1.3"
types-python-dateutil = "^2.8.19"
types-pytz = "^2022.2.1.0"
types-requests = "^2.28.10"
types-setuptools = "^65.3.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
pytest-cov = "^3.0.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--basetemp=testtemp"

[tool.coverage.run]
omit = ["*/test/*", "**/conftest.py"]

[tool.coverage.report]
omit = ["*/test/*"]
exclude_lines = [
    "pragma: no cover",
    "@abstractmethod",
    "@abc.abstractmethod",
    "def __repr__",
    "if .*DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_typing = "typing"
sections = [
    "FUTURE",
    "STDLIB",
    "FIRSTPARTY",
    "THIRDPARTY",
    "LOCALFOLDER",
    "TYPING",
]
# skip files
#skip_glob = docs/conf.py

[tool.mypy]
ignore_missing_imports = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
